[field_buffer] Move chunks logic to chunks file
